move -n/--send-nth-encounter checking into pokemon/_runner.py
doesn't belong into root runner
also store total runtime as int instead of float

+++ lib/pokemon/__init__.py
@@ -114,7 +114,7 @@ class PokemonRunner:
 
 	@property
 	@abstractmethod
-	def totalTime(self) -> float:
+	def totalTime(self) -> int:
 		raise NotImplementedError
 
 	@property
